api/requirements.txt
fastapi==0.110.2
idna==3.7
pydantic==2.7.1
#requests==2.7.0
requests==2.31.0
urllib3==2.2.1
uvicorn==0.29.0
typing==3.7.4.3
#scikit-learn==0.24.2
skforecast==0.11.0
pmdarima==2.0.4
numpy==1.26.4
pandas==2.1.4
python-binance==1.0.19
pymysql==1.1.0
sqlalchemy==2.0.29
mysql-connector-python==8.3.0
python-dotenv==1.0.1
lightgbm==4.3.0
ujson==5.9.0
